Organize the db dir
